[REFACTOR] 시간에 따른 운동기록 로직 리포지토리단으로 변경
