+++ src/main/java/com/dnd/ground/domain/exerciseRecord/service/ExerciseRecordService.java
@@ -9,7 +9,7 @@ import java.util.List;
  * @description 운동 기록 서비스 인터페이스
  * @author  박세헌
  * @since   2022-08-01
- * @updated 2022-08-01 / 생성 : 박세헌
+ * @updated 2022-08-01 / 운동기록로직 삭제: 박세헌
  */
 
 public interface ExerciseRecordService {
@@ -19,15 +19,4 @@ public interface ExerciseRecordService {
     ExerciseRecord findById(Long id);
 
     void delete(Long exerciseRecordId);
-
-    List<ExerciseRecord> findRecordOfThisWeek(Long id);
-
-    List<ExerciseRecord> findRecordOfPastByDay(Long id, LocalDateTime start);
-
-    List<ExerciseRecord> findRecordOfPastByWeek(Long id, LocalDateTime start);
-
-    List<ExerciseRecord> findChallengeRecordOfThisWeek(Long id, LocalDateTime start);
-
-    List<ExerciseRecord> findChallengeRecordOfPast(Long id, LocalDateTime start);
-
 }
